Change enterUncooperativeCode prototype to fix build on 32bit.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@142838 91177308-0d34-0410-b5e6-96231b3b80d8

+++ include/mvm/Threads/Thread.h
@@ -208,7 +208,7 @@ public:
   void joinRVBeforeEnter();
   void joinRVAfterLeave(word_t savedSP);
 
-  void enterUncooperativeCode(unsigned level = 0) __attribute__ ((noinline));
+  void enterUncooperativeCode(uint16_t level = 0) __attribute__ ((noinline));
   void enterUncooperativeCode(word_t SP);
   void leaveUncooperativeCode();
   word_t waitOnSP();
